Split PageTracker into logical and layout components, and added stories
